Prepare 0.5.8 release
