+++ apps/example/package.json
@@ -20,20 +20,20 @@
     "@babel/runtime": "^7.29.2",
     "@expo-google-fonts/inter": "^0.4.2",
     "@expo-google-fonts/jetbrains-mono": "^0.4.1",
-    "expo": "~56.0.9",
+    "expo": "~56.0.11",
     "expo-asset": "~56.0.16",
-    "expo-build-properties": "~56.0.17",
+    "expo-build-properties": "~56.0.18",
     "expo-constants": "~56.0.17",
     "expo-font": "~56.0.5",
     "expo-linking": "~56.0.13",
-    "expo-router": "~56.2.9",
+    "expo-router": "~56.2.10",
     "expo-status-bar": "~56.0.4",
     "expo-system-ui": "~56.0.5",
     "react": "19.2.3",
     "react-dom": "19.2.3",
     "react-native": "0.85.3",
     "react-native-nitro-modules": "0.35.7",
-    "react-native-nitro-storage": "*",
+    "react-native-nitro-storage": "workspace:*",
     "react-native-safe-area-context": "~5.7.0",
     "react-native-screens": "4.25.2",
     "react-native-web": "^0.21.2"
