+++ package.json
@@ -6,32 +6,31 @@
     "packages/*"
   ],
   "scripts": {
-    "setup": "node scripts/setup.js",
-    "publish-package": "node scripts/publish.js",
-    "publish-package:dry-run": "node scripts/publish.js --dry-run --skip-checks --yes",
-    "publish-package:dry": "node scripts/publish.js --dry-run --skip-checks --yes",
+    "setup": "bun scripts/setup.js",
+    "publish-package": "bun scripts/publish.js",
+    "publish-package:dry-run": "bun scripts/publish.js --dry-run --skip-checks --yes",
     "dev": "bun run example:start",
-    "build": "turbo run build",
-    "lint": "turbo run lint",
-    "lint:fix": "turbo run lint:fix",
-    "lint:check": "turbo run lint:check",
-    "format": "turbo run format",
-    "format:check": "turbo run format:check",
-    "clean": "turbo run clean",
-    "codegen": "turbo run codegen",
-    "typecheck": "turbo run typecheck",
-    "test:types": "turbo run test:types --filter=react-native-nitro-storage",
-    "test": "turbo run test",
-    "test:coverage": "turbo run test:coverage --filter=react-native-nitro-storage",
-    "benchmark": "turbo run benchmark --filter=react-native-nitro-storage",
-    "test:cpp": "turbo run test:cpp --filter=react-native-nitro-storage",
-    "test:cpp:asan": "turbo run test:cpp:asan --filter=react-native-nitro-storage",
-    "test:cpp:coverage": "turbo run test:cpp:coverage --filter=react-native-nitro-storage",
-    "test:cpp:tsan": "turbo run test:cpp:tsan --filter=react-native-nitro-storage",
-    "test:cpp:ubsan": "turbo run test:cpp:ubsan --filter=react-native-nitro-storage",
-    "check": "bun run lint:check -- --filter=react-native-nitro-storage && bun run format:check -- --filter=react-native-nitro-storage && bun run typecheck -- --filter=react-native-nitro-storage && bun run test:types -- --filter=react-native-nitro-storage && bun run test -- --filter=react-native-nitro-storage && bun run test:cpp -- --filter=react-native-nitro-storage",
-    "check:ci": "bun run check && bun run benchmark -- --filter=react-native-nitro-storage",
-    "audit:package": "bun run --cwd packages/react-native-nitro-storage check:pack && bun pm pack --cwd packages/react-native-nitro-storage --dry-run",
+    "build": "bun run --cwd packages/react-native-nitro-storage build",
+    "lint": "bun run --cwd packages/react-native-nitro-storage lint",
+    "lint:fix": "bun run --cwd packages/react-native-nitro-storage lint:fix",
+    "lint:check": "bun run --cwd packages/react-native-nitro-storage lint:check",
+    "format": "bun run --cwd packages/react-native-nitro-storage format",
+    "format:check": "bun run --cwd packages/react-native-nitro-storage format:check",
+    "clean": "bun run --cwd packages/react-native-nitro-storage clean",
+    "codegen": "bun run --cwd packages/react-native-nitro-storage codegen",
+    "typecheck": "bun run --cwd packages/react-native-nitro-storage typecheck",
+    "test:types": "bun run --cwd packages/react-native-nitro-storage test:types",
+    "test": "bun run --cwd packages/react-native-nitro-storage test",
+    "test:coverage": "bun run --cwd packages/react-native-nitro-storage test:coverage",
+    "benchmark": "bun run build && bun run --cwd packages/react-native-nitro-storage benchmark",
+    "test:cpp": "bun run --cwd packages/react-native-nitro-storage test:cpp",
+    "test:cpp:asan": "bun run --cwd packages/react-native-nitro-storage test:cpp:asan",
+    "test:cpp:coverage": "bun run --cwd packages/react-native-nitro-storage test:cpp:coverage",
+    "test:cpp:tsan": "bun run --cwd packages/react-native-nitro-storage test:cpp:tsan",
+    "test:cpp:ubsan": "bun run --cwd packages/react-native-nitro-storage test:cpp:ubsan",
+    "check": "bun run lint:check && bun run format:check && bun run typecheck && bun run test:types && bun run test && bun run test:cpp",
+    "check:ci": "bun run check && bun run benchmark",
+    "audit:package": "bun run codegen && bun run build && bun run --cwd packages/react-native-nitro-storage check:pack && bun pm pack --cwd packages/react-native-nitro-storage --dry-run",
     "release:preflight": "bun run check:ci && bun run example:check && bun run audit:package && bun run publish-package:dry-run",
     "example:start": "bun run --cwd apps/example start",
     "example:ios": "bun run --cwd apps/example ios",
@@ -63,7 +62,6 @@
     "react-test-renderer": "19.2.3",
     "rimraf": "^6.1.3",
     "ts-jest": "^29.4.11",
-    "turbo": "^2.9.14",
     "typescript": "6.0.3"
   },
   "overrides": {
@@ -72,10 +70,10 @@
     "react-native": "0.85.3",
     "react-native-nitro-modules": "0.35.7",
     "@types/react": "~19.2.14",
-    "expo": "~56.0.9",
+    "expo": "~56.0.11",
     "react-native-screens": "4.25.2",
-    "expo-constants": "~56.0.17",
-    "expo-linking": "~56.0.13"
+    "expo-constants": "~56.0.18",
+    "expo-linking": "~56.0.14"
   },
   "packageManager": "bun@1.3.14",
   "trustedDependencies": [
