+++ packages/react-native-nitro-storage/tsconfig.json
@@ -3,12 +3,27 @@
   "compilerOptions": {
     "outDir": "./lib/typescript",
     "rootDir": "./src",
-    "lib": ["ES2020", "DOM"],
-    "types": ["react-native"],
+    "lib": [
+      "ES2020",
+      "DOM"
+    ],
+    "types": [
+      "react-native"
+    ],
     "noUncheckedIndexedAccess": true,
     "exactOptionalPropertyTypes": true,
-    "useUnknownInCatchVariables": true
+    "useUnknownInCatchVariables": true,
+    "noImplicitOverride": true,
+    "noImplicitReturns": true,
+    "noFallthroughCasesInSwitch": true
   },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "lib", "nitrogen", "src/__tests__"]
+  "include": [
+    "src/**/*"
+  ],
+  "exclude": [
+    "node_modules",
+    "lib",
+    "nitrogen",
+    "src/__tests__"
+  ]
 }
