--- turbo.json
@@ -1,52 +1,0 @@
-{
-  "$schema": "https://turbo.build/schema.json",
-  "tasks": {
-    "build": {
-      "dependsOn": ["^build", "codegen"],
-      "outputs": ["lib/**", "dist/**"]
-    },
-    "lint": {
-      "dependsOn": ["^build"]
-    },
-    "lint:fix": {
-      "dependsOn": ["^build"]
-    },
-    "lint:check": {
-      "dependsOn": ["^build"]
-    },
-    "format": {},
-    "format:check": {},
-    "clean": {
-      "cache": false
-    },
-    "codegen": {
-      "inputs": ["src/**/*.nitro.ts", "nitro.json", "package.json"],
-      "outputs": ["nitrogen/generated/**"]
-    },
-    "typecheck": {},
-    "test:types": {},
-    "test": {},
-    "test:coverage": {
-      "outputs": ["coverage/**"]
-    },
-    "benchmark": {
-      "dependsOn": ["build"],
-      "cache": false
-    },
-    "test:cpp": {
-      "cache": false
-    },
-    "test:cpp:asan": {
-      "cache": false
-    },
-    "test:cpp:coverage": {
-      "cache": false
-    },
-    "test:cpp:tsan": {
-      "cache": false
-    },
-    "test:cpp:ubsan": {
-      "cache": false
-    }
-  }
-}
